requirements.txt
fastapi
uvicorn
pydantic
faker
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt==3.2.2
itsdangerous
